Moved resolve_hosts.rs out of flags.rs following the pattern of resolve_addr
