+++ cli/main.rs
@@ -52,6 +52,7 @@ pub mod ops;
 pub mod permissions;
 mod repl;
 pub mod resolve_addr;
+pub mod resolve_hosts;
 pub mod signal;
 pub mod source_maps;
 mod startup_data;
